[MOV] html_editor,web_editor: move models to html_editor
Since the web_editor module is deprecated and planned for deletion,
we move the models from web_editor to the html_editor module.
