+++ addons/html_editor/__manifest__.py
@@ -14,6 +14,9 @@ This addon provides an extensible, maintainable editor.
     'version': '1.0',
     'category': 'Hidden',
     'depends': ['base', 'bus', 'web'],
+    'data': [
+        'security/ir.model.access.csv',
+    ],
     'auto_install': True,
     'assets': {
         'web.assets_frontend': [
