+++ addons/html_editor/models/__init__.py
@@ -1,4 +1,10 @@
 from . import ir_attachment
 from . import ir_http
+from . import ir_qweb_fields
+from . import ir_ui_view
+from . import ir_websocket
+
+from . import models
+from . import test_models
 
 from . import html_field_history_mixin
