+++ addons/html_editor/models/ir_qweb_fields.py
