+++ addons/html_editor/models/ir_ui_view.py
