+++ addons/html_editor/models/ir_websocket.py
