+++ addons/web_editor/models/__init__.py
@@ -1,13 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import ir_qweb
-from . import ir_qweb_fields
-from . import ir_ui_view
-from . import ir_http
-from . import ir_websocket
-from . import models
-
 from . import assets
-
-from . import test_models
+from . import ir_qweb
